Refactor documentation generation process
### What's done:
* Added another generation after merge
