+++ info/buildSrc/src/main/kotlin/org/cqfn/diktat/generation/docs/GenerationAvailableRules.kt
@@ -1,33 +1,23 @@
-/**
- * This script paste available.rules.md and rules-mapping.md files to automatically generate table for White paper
- */
-
-@file:Suppress("FILE_NAME_MATCH_CLASS")
-
-package org.cqfn.diktat.ruleset.generation
-
-import org.cqfn.diktat.ruleset.utils.AUTO_END
-import org.cqfn.diktat.ruleset.utils.AUTO_TABLE
+package org.cqfn.diktat.generation.docs
 
 import java.io.File
 
-fun main() {
-    generateAvailableRules()
-}
-
+/**
+ * This function parses available-rules.md and rules-mapping.md files to automatically generate table for White paper
+ */
 @Suppress("MagicNumber", "UnsafeCallOnNullableType")
-private fun generateAvailableRules() {
-    val ruleMap = File("info/rules-mapping.md").readLines()
+fun generateAvailableRules(rootDir: File, wpDir: File) {
+    val ruleMap = File(rootDir, "rules-mapping.md").readLines()
         .drop(2)
         .map { it.drop(1).dropLast(1).split("|") }
         .map { RuleDescription(it[0].replace("\\s+".toRegex(), ""), it[1], it[2]) }
         .associateBy { it.ruleName }
-    File("info/available-rules.md").readLines()
+    File(rootDir, "available-rules.md").readLines()
         .drop(2)
         .map { it.drop(1).dropLast(1).split("|") }
         .map { it[2].replace("\\s+".toRegex(), "") to it[5] }
         .forEach { ruleMap[it.first]!!.config = it.second}
-    val newText = File("wp/sections/appendix.tex").readLines().toMutableList()
+    val newText = File(wpDir, "sections/appendix.tex").readLines().toMutableList()
     newText.removeAll(newText.subList(newText.indexOf("\\section*{available-rules}") + 1, newText.indexOf("\\section*{\\textbf{Diktat Coding Convention}}")))
     var index = newText.indexOf("\\section*{available-rules}") + 1
     AUTO_TABLE.trimIndent().lines().forEach { newText.add(index++, it) }
@@ -35,7 +25,7 @@ private fun generateAvailableRules() {
         .map { "${it.correctRuleName} & ${it.correctCodeStyle} & ${it.autoFix} & ${it.config.replace("<br>", " ")}\\\\" }
         .forEach { newText.add(index++, it) }
     AUTO_END.trimIndent().split("\n").forEach { newText.add(index++, it) }
-    File("wp/sections/appendix.tex").writeText(newText.joinToString(separator = "\n"))
+    File(wpDir, "sections/appendix.tex").writeText(newText.joinToString(separator = "\n"))
 }
 
 /**
